Added csv writing capabilities as well as cache checking. Next is bug fixing and testing

+++ XRootESClient/xroot_es_client_jsonlines.py
@@ -7,6 +7,7 @@
 DEBUGSITE=0
 import json
 import jsonlines
+import csv
 import os
 import sys
 import time
@@ -422,6 +423,11 @@ class XRootESClient():
             is_start = True
             if self.args["display_aggregate"]:
                 aggregates = {}
+
+        if self.args["make_csvs"]:
+            csv_filename = f"{self.dirname}/{self.args['experiment']}_summary_{self.args['start_date']}_{self.args['end_date']}.csv"
+            csv_file = open(csv_filename, 'w')
+            csv_writer = None
         #Steps through all found project IDs
         for pid in self.pid_list:
             #Checks to make sure that the raw file we're accessing isn't empty
@@ -430,6 +436,7 @@ class XRootESClient():
                 if self.args["clear_raws"]:
                     os.remove(self.pids[pid]["raw_filename"])
                 continue
+
             #Each PID has an associated raw file, with a saved name for convenience
             with jsonlines.open(self.pids[pid]["raw_filename"]) as reader:
                 #Sets some default values for checks or inter-iteration variables
@@ -520,7 +527,12 @@ class XRootESClient():
                                             aggregates[pid][to_write["source"]][to_write["destination"]]["data_tier"].append(to_write["data_tier"])
                                         aggregates[pid][to_write["source"]][to_write["destination"]]["file_count"] += 1
 
-
+                            if self.args["make_csvs"]:
+                                if csv_writer is None:
+                                    fieldnames = summary.keys()
+                                    csv_writer = csv.DictWriter(csv_file, fieldnames=fieldnames)
+                                    csv_writer.writeheader()
+                                csv_writer.writerow(summary)
 
                         #Updates the value of the last checked event and event
                         #number to the current event and event number
@@ -621,6 +633,10 @@ class XRootESClient():
                             display_writer.write(json.dumps(info_out, indent=2))
             display_writer.write("\n]}")
             display_writer.close()
+
+        if self.args["make_csvs"]:
+            csv_writer.close()
+
         sum_writer.close()
 
         #Gets summarizer end time and duration
@@ -639,8 +655,10 @@ class XRootESClient():
         #Makes all of the raw file objects
         for pid in self.pid_list:
             fname = f"{self.dirname}/{self.args['experiment']}_raw_{self.args['start_date']}_{self.args['end_date']}_{pid}.jsonl"
-            self.pids[pid]["raw_filename"] = fname
-            proj_files[pid] = open(fname, "w+")
+            if self.args["overwrite_files"] or os.path.exists(fname):
+                self.pids[pid]["raw_filename"] = fname
+                proj_files[pid] = open(fname, "w+")
+
 
         #Checks that either the compiler is still running, or there's still data in
         #the finished data queue to prevent early shutdowns
@@ -720,15 +738,22 @@ class XRootESClient():
                     print(f"Making compiler thread for pid {self.pid_list[i]}")
                 pid = self.pid_list[i]
                 #Decrements the semaphore upon spawning a new thread
-                self.inc_lock.acquire()
-                self.max_compiler_threads -= 1
-                self.inc_lock.release()
-                self.pids[pid]["compiler_worker"] = threading.Thread(target=self.compiler_worker_func,args=(pid,),daemon=True)
-                self.pids[pid]["compiler_worker"].start()
+                raw_fname = f"{self.dirname}/{self.args['experiment']}_raw_{self.args['start_date']}_{self.args['end_date']}_{pid}.jsonl"
+                if self.args["overwrite_files"] or os.path.exists(raw_fname):
+                    self.inc_lock.acquire()
+                    self.max_compiler_threads -= 1
+                    self.inc_lock.release()
+                    self.pids[pid]["compiler_worker"] = threading.Thread(target=self.compiler_worker_func,args=(pid,),daemon=True)
+                    self.pids[pid]["compiler_worker"].start()
+                else:
+                    self.inc_lock.acquire()
+                    self.finished_compiler_threads += 1
+                    self.inc_lock.release()
                 i += 1
         #Waits for all threads to terminate
         for pid in self.pid_list:
-            self.pids[pid]["compiler_worker"].join()
+            if not self.pids[pid]["compiler_worker"] is None:
+                self.pids[pid]["compiler_worker"].join()
 
     #Compiles all data into finished form for a given project
     def compiler_worker_func(self, pid):
@@ -840,18 +865,25 @@ class XRootESClient():
                 if self.debug > 3:
                     print(f"Making SAM thread for pid {self.pid_list[i]}")
                 pid = self.pid_list[i]
-                #Decrements the semaphore when spawning a new thread
-                self.inc_lock.acquire()
-                self.max_sam_threads -= 1
-                self.inc_lock.release()
-                self.pids[pid]["sam_worker"] = threading.Thread(target=self.sam_worker_func,args=(pid,),daemon=True)
-                self.pids[pid]["sam_worker"].start()
-                self.total_pids += 1
+                raw_fname = f"{self.dirname}/{self.args['experiment']}_raw_{self.args['start_date']}_{self.args['end_date']}_{pid}.jsonl"
+                if self.args["overwrite_files"] or os.path.exists(raw_fname):
+                    #Decrements the semaphore when spawning a new thread
+                    self.inc_lock.acquire()
+                    self.max_sam_threads -= 1
+                    self.inc_lock.release()
+                    self.pids[pid]["sam_worker"] = threading.Thread(target=self.sam_worker_func,args=(pid,),daemon=True)
+                    self.pids[pid]["sam_worker"].start()
+                    self.total_pids += 1
+                else:
+                    self.inc_lock.acquire()
+                    self.finished_sam_threads += 1
+                    self.inc_lock.release()
                 i += 1
 
         #Waits for all threads to terminate
         for pid in self.pid_list:
-            self.pids[pid]["sam_worker"].join()
+            if not self.pids[pid]["sam_worker"] is None:
+                self.pids[pid]["sam_worker"].join()
 
     def sam_worker_func(self, pid):
         #Runs as long as there's data, or the thread that produces data is either alive or hasn't been started
@@ -902,16 +934,23 @@ class XRootESClient():
                 pid = self.pid_list[i]
                 if self.debug > 3:
                     print(f"Making ElasticSearch thread for pid {pid}")
-                #Decrements the sempahore when spawning a new thread
-                self.inc_lock.acquire()
-                self.max_es_threads -= 1
-                self.inc_lock.release()
-                self.pids[pid]["es_worker"] = threading.Thread(target=self.es_worker_func,args=(pid,),daemon=True)
-                self.pids[pid]["es_worker"].start()
+                raw_fname = f"{self.dirname}/{self.args['experiment']}_raw_{self.args['start_date']}_{self.args['end_date']}_{pid}.jsonl"
+                if self.args["overwrite_files"] or os.path.exists(raw_fname):
+                    #Decrements the sempahore when spawning a new thread
+                    self.inc_lock.acquire()
+                    self.max_es_threads -= 1
+                    self.inc_lock.release()
+                    self.pids[pid]["es_worker"] = threading.Thread(target=self.es_worker_func,args=(pid,),daemon=True)
+                    self.pids[pid]["es_worker"].start()
+                else:
+                    self.inc_lock.acquire()
+                    self.finished_es_threads += 1
+                    self.inc_lock.release()
                 i += 1
         #Waits for all threads to terminate
         for pid in self.pid_list:
-            self.pids[pid]["es_worker"].join()
+            if not self.pids[pid]["es_worker"] is None:
+                self.pids[pid]["es_worker"].join()
 
     #Function to pull all logged SAM Events for a given PID in the established date range
     def es_worker_func(self, pid):
